+++ frontend/src/toolbar/actions/actionsLogic.js
@@ -2,7 +2,7 @@ import { kea } from 'kea'
 import { toolbarLogic } from '~/toolbar/toolbarLogic'
 import { encodeParams } from 'kea-router'
 import { currentPageLogic } from '~/toolbar/stats/currentPageLogic'
-import { stepMatchesHref } from '~/toolbar/elements/utils'
+import { stepMatchesHref } from '~/toolbar/utils'
 
 export const actionsLogic = kea({
     loaders: ({ values }) => ({
